* tag 'firefox_84_release': (17258 commits)
  No bug - Tagging 98507f230e7442065b287ced065ac77d4d8ea35d with FIREFOX_84_0_RELEASE a=release CLOSED TREE DONTBUILD
  No Bug, mozilla-release repo-update remote-settings - a=repo-update r=jcristau
  No bug - Tagging 98507f230e7442065b287ced065ac77d4d8ea35d with FIREFOX_84_0_BUILD3 a=release CLOSED TREE DONTBUILD
  Bug 1681328 - add more domains for Amazon and eBay tippytop icons. r=mikedeboer, a=RyanVM
  No bug - Tagging c513faba0ecb007328f05db3a98e3ba1aba7f10f with FIREFOX_84_0_BUILD2 a=release CLOSED TREE DONTBUILD
  Bug 1681428 - back out highlights and logo changes. r=thecount, a=RyanVM
  Bug 1676474 - Disable webrender partial present on Mali-Gxx. r=aosmond, a=RyanVM
  Bug 1681251 - Return AUTF8String from GetDateTimePattern in mozIOSPreferences; r=zbraniecki, a=RyanVM
  No Bug, mozilla-release repo-update remote-settings - a=repo-update r=RyanVM
  Bug 1680528 - set several webrtc/mediacapturing test as passing on beta and release. r=drno a=test-only
  Bug 1677797 - Change autoconf to avoid using a hard-coded file descriptor. r=glandium, a=RyanVM
  No bug - Tagging 6cf020352b4e350ab6b84253620652d55cdd0b60 with FIREFOX_84_0_BUILD1 a=release CLOSED TREE DONTBUILD
  Backed out changeset 346a46a93b30 (bug 1680299) to have universal builds. a=RyanVM
  Bug 1680299 - Turn the mac shippable build back into an x64_64-only build. r=glandium, a=RyanVM
  Bug 1681022 - Ignore flex-basis: content on a table wrapper. r=dholbert, a=RyanVM
  Update configs. IGNORE BROKEN CHANGESETS CLOSED TREE NO BUG a=release ba=release
  No bug - tagging a42ea3aab3e467141d56b5c299749ee842cd3c41 with FIREFOX_RELEASE_83_END a=release DONTBUILD CLOSED TREE
  Preserve old tags after debusetparents. CLOSED TREE DONTBUILD a=release
  No bug - tagging 87ab4da1b9f786db1f5c6633357cbe2ee1ca562e with FIREFOX_RELEASE_84_BASE a=release DONTBUILD CLOSED TREE
  no bug - Bumping Firefox l10n changesets r=release a=l10n-bump
  ...

+++ python/mozperftest/mozperftest/system/proxy.py
@@ -5,11 +5,12 @@ import json
 import os
 import re
 import signal
+import tempfile
 import threading
 
 from mozlog import get_proxy_logger
 from mozperftest.layers import Layer
-from mozperftest.utils import install_package
+from mozperftest.utils import download_file, install_package
 from mozprocess import ProcessHandler
 
 
@@ -24,10 +25,10 @@ class OutputHandler(object):
         self.port_event = threading.Event()
 
     def __call__(self, line):
-        if not line.strip():
+        line = line.strip()
+        if not line:
             return
         line = line.decode("utf-8", errors="replace")
-
         try:
             data = json.loads(line)
         except ValueError:
@@ -39,7 +40,7 @@ class OutputHandler(object):
             # our subprocess.
             m = re.match(r"Proxy running on port (\d+)", data.get("message", ""))
             if m:
-                self.port = m.group(1)
+                self.port = int(m.group(1))
                 self.port_event.set()
             LOG.log_raw(data)
         else:
@@ -49,7 +50,10 @@ class OutputHandler(object):
         self.port_event.set()
 
     def process_output(self, line):
-        LOG.process_output(self.proc.pid, line)
+        if self.proc is None:
+            LOG.process_output(line)
+        else:
+            LOG.process_output(self.proc.pid, line)
 
     def wait_for_port(self):
         self.port_event.wait()
@@ -57,15 +61,28 @@ class OutputHandler(object):
 
 
 class ProxyRunner(Layer):
-    """Use a proxy
-    """
+    """Use a proxy"""
 
     name = "proxy"
     activated = False
 
+    arguments = {
+        "record": {
+            "type": str,
+            "default": None,
+            "help": "Generate a recording of the network requests during this test.",
+        },
+        "replay": {
+            "type": str,
+            "default": None,
+            "help": "A generated recording to play back during this test.",
+        },
+    }
+
     def __init__(self, env, mach_cmd):
         super(ProxyRunner, self).__init__(env, mach_cmd)
         self.proxy = None
+        self.tmpdir = None
 
     def setup(self):
         # Install mozproxy and its vendored deps.
@@ -76,19 +93,34 @@ class ProxyRunner(Layer):
 
     def run(self, metadata):
         self.metadata = metadata
+        replay_file = self.get_arg("replay")
+        if replay_file is not None and replay_file.startswith("http"):
+            self.tmpdir = tempfile.TemporaryDirectory()
+            target = os.path.join(self.tmpdir.name, "recording.zip")
+            self.info("Downloading %s" % replay_file)
+            download_file(replay_file, target)
+            replay_file = target
 
         self.info("Setting up the proxy")
-        # replace with artifacts
+        command = [
+            self.mach_cmd.virtualenv_manager.python_path,
+            "-m",
+            "mozproxy.driver",
+            "--local",
+            "--binary=" + self.mach_cmd.get_binary_path(),
+            "--topsrcdir=" + self.mach_cmd.topsrcdir,
+            "--objdir=" + self.mach_cmd.topobjdir,
+        ]
+        if self.get_arg("record"):
+            command.extend(["--record", self.get_arg("record")])
+        elif replay_file:
+            command.append(replay_file)
+        else:
+            command.append(os.path.join(HERE, "example.zip"))
+        print(" ".join(command))
         self.output_handler = OutputHandler()
         self.proxy = ProcessHandler(
-            [
-                "mozproxy",
-                "--local",
-                "--binary=" + self.mach_cmd.get_binary_path(),
-                "--topsrcdir=" + self.mach_cmd.topsrcdir,
-                "--objdir=" + self.mach_cmd.topobjdir,
-                os.path.join(HERE, "example.dump"),
-            ],
+            command,
             processOutputLine=self.output_handler,
             onFinish=self.output_handler.finished,
         )
@@ -115,8 +147,21 @@ class ProxyRunner(Layer):
         return metadata
 
     def teardown(self):
+        err = None
         if self.proxy is not None:
-            kill_signal = getattr(signal, "CTRL_BREAK_EVENT", signal.SIGINT)
-            os.kill(self.proxy.pid, kill_signal)
-            self.proxy.wait()
+            returncode = self.proxy.wait(0)
+            if returncode is not None:
+                err = ValueError(
+                    "mozproxy terminated early with return code %d" % returncode
+                )
+            else:
+                kill_signal = getattr(signal, "CTRL_BREAK_EVENT", signal.SIGINT)
+                os.kill(self.proxy.pid, kill_signal)
+                self.proxy.wait()
             self.proxy = None
+        if self.tmpdir is not None:
+            self.tmpdir.cleanup()
+            self.tmpdir = None
+
+        if err:
+            raise err
